feat: v0.197.0 -- rcp_byte_bus_id_t widened to the full 11-bit wire range, Group 3 fully complete (REQ-RMAP-053/REQ-ACF-020, issue #200)
Full "read every consumer" investigation performed first (~55 files
touching rcp_byte_bus_id_t), per this item's own standing flag from
earlier batches. rcp_byte_bus_id_t (avtp.h) widened uint8_t ->
uint16_t, wide enough for TC18's own 11-bit BBID field.

Every endpoint type module, the conditional-request layer, config.c,
lifecycle.c, mock.c, and server.c were confirmed to be simple
pass-through consumers needing no logic changes. acf.c's encode path
was already written generically for 11 bits (the bug was purely the
typedef's own width, not a second encode bug); the decode path was
already honestly rejecting out-of-range values rather than corrupting
them. Removed the now-provably-dead overflow check in
rcp_acf_unpack_header() (busid_full is mathematically bounded to
0x7FF) and retired RCP_ACF_ERR_BUS_ID_OVERFLOW outright from
rcp_acf_errc_t (its own last value, nothing else referenced it,
checked errors.c's TC18 Table 27 mapping to be sure).

Two real consumer-side fixes found and made alongside: adapt.c's
byte_bus_id-to-string buffer widened (char[4] -> char[6], the old size
would have silently mis-formatted any id above 999); test_mock.c's own
decode-failure test given a new, still-valid stimulus (an
over-declared pad count) since its old one (a high byte_bus_id) no
longer fails to decode.

Rewrote three width/reject-pinning deviation tests into positive
round-trip tests; split one combined test that used to also cover
REQ-ACF-018's own unrelated, still-open deviation. Renamed one new
test after cfusa check flagged a CFUSA-CY009 substring false positive
("decodes_full" matching the DES-cipher "des_" pattern).

Mutation-tested with three independent mutations (full typedef
revert; isolated decode-truncation; isolated encode-masking) -- all
caught by the same three tests from different angles. Full suite
(65/65, net +2 tests) + ASan/UBSan clean on both trees. Fresh cfusa
check (0 errors) + all three trace invocations (100%/100%).
REQ-RMAP-053 and REQ-ACF-020 both move to implemented.

Group 3 (EP_ID_config, §12.7.8 Table 23) is now COMPLETE: 6/6 items.

+++ include/rcp/avtp.h
@@ -123,8 +123,14 @@ typedef struct {
 
 /* Endpoint-within-stream address. Unique only within the stream_id it is
  * paired with (see rcp_avtp_addr_t) -- never globally, and never on its
- * own. */
-typedef uint8_t rcp_byte_bus_id_t;
+ * own. TC18's own BBID field is 11 bits wide (0-2047, ACF byte_message_
+ * info octet 2 bits [2:0] carrying [10:8] and octet 3 carrying [7:0],
+ * see acf.h; also the EP_ID_config table's own BBID register, regmap.h)
+ * -- uint16_t comfortably holds the full range (REQ-RMAP-053/REQ-ACF-020,
+ * fixed as of this typedef's own follow-up batch; values above 2047 are
+ * simply never produced by any encoder/decoder in this codebase, not a
+ * distinct wire representation this type needs to reject on its own). */
+typedef uint16_t rcp_byte_bus_id_t;
 
 /* Builds a stream_id from a 6-byte MAC and a caller-assigned unique_id. */
 rcp_stream_id_t rcp_stream_id_make(const uint8_t mac[6], uint16_t unique_id);
